Refactoring the Slugger.
